Added support for empty objects
decode/encode is asymetrical now but can't fix that without completely changing how objects are received.

previously, it was impossible to create { key: {} }, since [] is interpreted as an empty string and there was no means to represent an empty proplist.

the asymmetry comes from a find(X) where the object is { data: {} } will return
[{data,[]}]
but encoding that object will look like:
{ data: "" }

you would have to store it with:
[{data,{obj,[]}}]

to get the empty map